fixes cythonization and refactors
